refact(ProductPageBlock): disable active section dectection for mobile
