--- apps/store/src/blocks/ProductPageBlock.tsx
@@ -1,331 +1,0 @@
-import { css, Global } from '@emotion/react'
-import styled from '@emotion/styled'
-import * as RadixTabs from '@radix-ui/react-tabs'
-import { storyblokEditable, StoryblokComponent, SbBlokData } from '@storyblok/react'
-import { motion, Transition } from 'framer-motion'
-import { useRef, useState, useEffect } from 'react'
-import { mq, theme } from 'ui'
-import {
-  MENU_BAR_HEIGHT_MOBILE,
-  MENU_BAR_HEIGHT_DESKTOP,
-  MENU_BAR_HEIGHT_PX,
-} from '@/components/Header/HeaderStyles'
-import { useProductPageContext } from '@/components/ProductPage/ProductPageContext'
-import { PurchaseForm } from '@/components/ProductPage/PurchaseForm/PurchaseForm'
-import * as Tabs from '@/components/ProductPage/Tabs'
-import { ProductVariantSelector } from '@/components/ProductVariantSelector/ProductVariantSelector'
-import { SbBaseBlockProps } from '@/services/storyblok/storyblok'
-import { useScrollState } from '@/utils/useScrollState'
-import { zIndexes } from '@/utils/zIndex'
-
-const TABLIST_HEIGHT = '2.5rem'
-
-// TODO: move to theme
-const TRANSITION: Transition = { ease: [0.65, 0.05, 0.36, 1] }
-
-type PageSection = 'overview' | 'coverage'
-
-type ProductPageBlockProps = SbBaseBlockProps<{
-  overviewLabel: string
-  coverageLabel: string
-  overview: SbBlokData[]
-  coverage: SbBlokData[]
-  body: SbBlokData[]
-}>
-
-export const ProductPageBlock = ({ blok }: ProductPageBlockProps) => {
-  const { productData } = useProductPageContext()
-  const [activeSection, setActiveSection] = useState<PageSection>('overview')
-
-  useActiveSectionChangeListener(['overview', 'coverage'], (sectionId) =>
-    setActiveSection(sectionId as PageSection),
-  )
-
-  const shouldRenderVariantSelector =
-    activeSection === 'coverage' && productData.variants.length > 1
-
-  return (
-    <>
-      <Global
-        styles={{
-          html: {
-            [mq.md]: {
-              scrollPaddingTop: MENU_BAR_HEIGHT_DESKTOP,
-            },
-          },
-        }}
-      />
-      <Main {...storyblokEditable(blok)}>
-        <MobileLayout>
-          <PurchaseForm />
-          <Content>
-            <RadixTabs.Tabs
-              value={activeSection}
-              onValueChange={(value) => {
-                setActiveSection(value as PageSection)
-                window?.scrollTo({ top: 0 })
-              }}
-            >
-              <AnimatedHeader>
-                <RadixTabs.TabsList>
-                  <TablistWrapper>
-                    <TabTrigger value="overview">{blok.overviewLabel}</TabTrigger>
-                    <TabTrigger value="coverage">{blok.coverageLabel}</TabTrigger>
-                  </TablistWrapper>
-                  {shouldRenderVariantSelector && <StyledProductVariantSelector />}
-                </RadixTabs.TabsList>
-              </AnimatedHeader>
-
-              <OverviewSection>
-                <Tabs.TabsContent value="overview">
-                  {blok.overview?.map((nestedBlock) => (
-                    <StoryblokComponent blok={nestedBlock} key={nestedBlock._uid} />
-                  ))}
-                </Tabs.TabsContent>
-              </OverviewSection>
-
-              <Tabs.TabsContent value="coverage">
-                {blok.coverage?.map((nestedBlock) => (
-                  <StoryblokComponent blok={nestedBlock} key={nestedBlock._uid} />
-                ))}
-              </Tabs.TabsContent>
-            </RadixTabs.Tabs>
-          </Content>
-        </MobileLayout>
-
-        <DesktopLayout>
-          <AnimatedHeader>
-            <nav aria-label="page content">
-              <ContentNavigationList>
-                <li>
-                  <ContentNavigationTrigger
-                    href="#overview"
-                    data-state={activeSection === 'overview' ? 'active' : 'inactive'}
-                    aria-current={activeSection === 'overview' ? 'true' : undefined}
-                  >
-                    {blok.overviewLabel}
-                  </ContentNavigationTrigger>
-                </li>
-                <li>
-                  <ContentNavigationTrigger
-                    href="#coverage"
-                    data-state={activeSection === 'coverage' ? 'active' : 'inactive'}
-                    aria-current={activeSection === 'coverage' ? 'true' : undefined}
-                  >
-                    {blok.coverageLabel}
-                  </ContentNavigationTrigger>
-                </li>
-              </ContentNavigationList>
-              {shouldRenderVariantSelector && <StyledProductVariantSelector />}
-            </nav>
-          </AnimatedHeader>
-          <Grid>
-            <Content>
-              <OverviewSection id="overview">
-                {blok.overview?.map((nestedBlock) => (
-                  <StoryblokComponent blok={nestedBlock} key={nestedBlock._uid} />
-                ))}
-              </OverviewSection>
-            </Content>
-            <PurchaseFormWrapper>
-              <PurchaseForm />
-            </PurchaseFormWrapper>
-          </Grid>
-          <section id="coverage">
-            {blok.coverage?.map((nestedBlock) => (
-              <StoryblokComponent blok={nestedBlock} key={nestedBlock._uid} />
-            ))}
-          </section>
-        </DesktopLayout>
-
-        {blok.body.map((nestedBlock) => (
-          <StoryblokComponent blok={nestedBlock} key={nestedBlock._uid} />
-        ))}
-      </Main>
-    </>
-  )
-}
-ProductPageBlock.blockName = 'product'
-
-const sharedListStyles = css({
-  display: 'flex',
-  gap: theme.space.xs,
-  height: TABLIST_HEIGHT,
-})
-
-const sharedTriggerStyles = css({
-  cursor: 'pointer',
-  display: 'flex',
-  alignItems: 'flex-end',
-  justifyContent: 'center',
-  textAlign: 'center',
-  paddingInline: theme.space.md,
-  paddingBlock: theme.space.xs,
-  fontSize: theme.fontSizes.md,
-  lineHeight: theme.fontSizes.xl,
-  color: theme.colors.dark,
-  backgroundColor: theme.colors.grayTranslucent100,
-  boxShadow: '0px 1px 2px rgba(0, 0, 0, 0.15)',
-  backdropFilter: 'blur(20px)',
-  borderRadius: theme.radius.sm,
-
-  '&[data-state=active]': {
-    paddingInline: theme.space.xxxl,
-    color: theme.colors.dark,
-    backgroundColor: theme.colors.grayTranslucent400,
-  },
-
-  '@media (hover: hover)': {
-    ':hover': {
-      backgroundColor: theme.colors.gray200,
-      transition: `background-color ${theme.transitions.hover}`,
-
-      '&[data-state=active]': {
-        backgroundColor: theme.colors.grayTranslucent500,
-      },
-    },
-  },
-
-  '&:focus-visible': {
-    boxShadow: `0 0 0 2px ${theme.colors.purple500}`,
-  },
-})
-
-const Main = styled.main({ width: '100%' })
-
-const PurchaseFormWrapper = styled.div({
-  position: 'sticky',
-  top: '9vh',
-  // Scroll independently if content is too long
-  maxHeight: '100vh',
-  overflow: 'auto',
-  paddingBottom: theme.space.xl,
-})
-
-const OverviewSection = styled.section({
-  marginTop: `calc(-${TABLIST_HEIGHT} - ${theme.space.xs})`,
-  [mq.md]: {
-    marginTop: `calc(-${TABLIST_HEIGHT} - ${theme.space.md})`,
-  },
-  [mq.lg]: {
-    marginTop: 0,
-  },
-})
-
-const MobileLayout = styled.div({
-  [mq.lg]: {
-    display: 'none',
-  },
-})
-
-const DesktopLayout = styled.div({
-  display: 'none',
-  [mq.lg]: {
-    display: 'block',
-  },
-})
-
-const Grid = styled.div({
-  display: 'grid',
-  gridTemplateColumns: 'repeat(2, 1fr)',
-  alignItems: 'flex-start',
-})
-
-const Content = styled.div({
-  position: 'relative',
-  display: 'flex',
-  flexDirection: 'column',
-  isolation: 'isolate',
-})
-
-const TablistWrapper = styled.div(sharedListStyles)
-
-const TabTrigger = styled(RadixTabs.Trigger)(sharedTriggerStyles)
-
-const AnimatedHeader = ({ children }: { children: React.ReactNode }) => {
-  const state = useScrollState({ threshold: MENU_BAR_HEIGHT_PX })
-  const scrollUp = state === 'SCROLL_DOWN' || state === 'BELOW'
-
-  return (
-    <StickyHeader animate={{ top: scrollUp ? theme.space.md : undefined }} transition={TRANSITION}>
-      {children}
-    </StickyHeader>
-  )
-}
-
-const StickyHeader = styled(motion.div)({
-  position: 'sticky',
-  top: `calc(${theme.space.sm} + ${MENU_BAR_HEIGHT_MOBILE})`,
-  zIndex: zIndexes.tabs,
-  paddingInline: theme.space.md,
-
-  [mq.md]: {
-    top: `calc(${theme.space.sm} + ${MENU_BAR_HEIGHT_DESKTOP})`,
-    paddingInline: theme.space.lg,
-  },
-  [mq.lg]: {
-    position: 'fixed',
-    top: `calc(${theme.space.md} + ${MENU_BAR_HEIGHT_DESKTOP})`,
-    paddingInline: theme.space.xl,
-  },
-})
-
-const ContentNavigationList = styled.ol(sharedListStyles)
-
-const ContentNavigationTrigger = styled.a(sharedTriggerStyles)
-
-const StyledProductVariantSelector = styled(ProductVariantSelector)({
-  minWidth: '12.5rem',
-  width: 'fit-content',
-  marginTop: theme.space.xs,
-})
-
-function useActiveSectionChangeListener(
-  sectionsId: Array<string>,
-  callback: (activeSectionId: string) => void,
-) {
-  // Used to determine if user is scrolling up
-  const scrollPositionRef = useRef(0)
-
-  useEffect(
-    function identifyActiveSection() {
-      const instersectionObserverCb = (entries: IntersectionObserverEntry[]) => {
-        entries.forEach((entry) => {
-          const id = entry.target.id
-          const diff = scrollPositionRef.current - window.scrollY
-          const isScrollingUp = diff > 0
-
-          if (isScrollingUp) {
-            if (!entry.isIntersecting) {
-              const activeSectionIndex = sectionsId.findIndex((sectionId) => sectionId === id)
-              const previousSectionId = sectionsId[activeSectionIndex - 1]
-              if (previousSectionId) {
-                callback(previousSectionId)
-                scrollPositionRef.current = window.scrollY
-              }
-            }
-          } else {
-            if (entry.isIntersecting) {
-              callback(id)
-              scrollPositionRef.current = window.scrollY
-            }
-          }
-        })
-      }
-
-      const observer = new IntersectionObserver(instersectionObserverCb, {
-        // Observe only the top 15% of the screen
-        rootMargin: '0% 0% -85% 0%',
-      })
-      sectionsId.forEach((sectionId) => {
-        const element = document.getElementById(sectionId)
-        if (element) {
-          observer.observe(element)
-        }
-      })
-
-      return () => observer.disconnect()
-    },
-    [sectionsId, callback],
-  )
-}
