+++ apps/store/src/blocks/ProductPageBlock/ProductPageBlock.types.tsx
@@ -1,0 +1,1 @@
+export type PageSection = 'overview' | 'coverage'
